2nd commit
Data analysis and visualization. 

Here, we have taken the large csv files and selected only the 'close' values to compare. 

We produced graphs for both the coinbase and bitstamp files and then overlayed them to see how similar or different the closing values are. We see there is very high correlation between the two.